Update RecipesListView.swift
This is the moneymaker - we need to be sure to set `newRecipe` to an actual new recipe everytime the user opens up the `+` in the toolbar to create a new recipe. This 1) creates a new empty recipes (since we created `emptyRecipe` to return a new recipe every time and 2) resets the forms to empty.