Update ModifyIngredientView.swift
This is more of an efficiency fix, maybe? 

To my understanding, a computed property is computed each and every time. Thus, `NumberFormatter` is computed for every case. Sure, in this application, it doesn't mean a super speed increase. But, would it be more efficient to use a `lazy var` computed property? The computation would only ever happen once. I've also attached an image of the same ideas with `Int`.